+++ src/lib/Details/UserDetails.svelte
@@ -1,42 +1,21 @@
 <script lang="ts">
-import { each } from 'svelte/internal';
-
-	// Importar por nombre de componentes: https://sveltestrap.js.org/
-	import {
-		Button,
-		Breadcrumb,
-		BreadcrumbItem,
-		Card,
-		CardBody,
-		CardFooter,
-		CardHeader,
-		CardSubtitle,
-		CardText,
-		CardTitle,
-		TabContent,
-		TabPane,
-		Image
-	} from 'sveltestrap';
-
-	// Info usuario placeholder (esto lo recibe del servidor en estructura similar):
-	let userDetails = {
-		user_id: '1234',
-		cuit: '20301001008',
-		name: 'Juan',
-		lastName: 'Perez',
-		email: 'juan.perez@ejemplo.com',
-		phone: '2993334444',
-		gender: 'M',
-		dateOfBirth: '1980-12-31',
-		nationality: 'Argentina',
-		studyLevel: 'Universitario completo',
-		degree: 'Licenciado',
-		profilePic: 'https://avatars.dicebear.com/api/micah/1234.svg',
-		roles: [
-			{ rol_id: 1, rolDescription: 'Gestor documental' },
-			{ rol_id: 2, rolDescription: 'Personal de seguridad' }
-		]
+	// Datos del usuario a mostrar
+	export let userDetails = {
+		user_id: '',
+		cuit: '',
+		name: '',
+		lastName: '',
+		email: '',
+		phone: '',
+		gender: '',
+		dateOfBirth: '',
+		nationality: '',
+		studyLevel: '',
+		degree: '',
+		profilePic: '',
+		roles: []
 	};
+
 	// Pasar inicial de género a texto legible:
 	let genderText;
 	if (userDetails.gender == 'M') {
@@ -48,44 +27,6 @@ import { each } from 'svelte/internal';
 	}
 </script>
 
-<svelte:head>
-	<!-- Insertar al head del HTML -->
-	<title>Detalles: {userDetails.name + ' ' + userDetails.lastName} - SeguCheck</title>
-</svelte:head>
-
-<!-- Encabezado -->
-<header class="row">
-	<Breadcrumb>
-		<BreadcrumbItem>
-			<a href="/panel/">Inicio</a>
-		</BreadcrumbItem>
-		<BreadcrumbItem>
-			<a href="/panel/usuarios">Usuarios</a>
-		</BreadcrumbItem>
-		<BreadcrumbItem active>Detalles</BreadcrumbItem>
-	</Breadcrumb>
-	<div class="col-md-3">
-		<Image
-			fluid
-			thumbnail
-			src={userDetails.profilePic}
-			alt="Aplicación móvil"
-			class="m-2"
-			style="max-width:150px"
-		/>
-	</div>
-	<div class="col-auto">
-		<h1>{userDetails.name + ' ' + userDetails.lastName}</h1>
-		<p class="lead">Detalles del usuario</p>
-	</div>
-	<div class="col-2 ms-auto">
-		<Button color="primary" href="/panel/operarios/nuevo"
-			><i class="fas fa-pen me-2" />Editar</Button
-		>
-	</div>
-</header>
-
-<!-- Detalles -->
 <form name="formDetails" id="formDetails">
 	<div class="row mb-3 g-3">
 		<div class="col-md-6">
@@ -249,14 +190,16 @@ import { each } from 'svelte/internal';
 				class="form-control overflow-visible"
 				aria-label="Roles"
 				value={userDetails.degree}
-				multiple>
-			{#if (userDetails.roles.length === 0)}
-				<option selected>Ninguno</option>
-			{:else}
-				{#each userDetails.roles as {rol_id, rolDescription}}
-					<option value="{rol_id}">{rolDescription}</option>
-				{/each}
-			{/if}
+				multiple
+			>
+				{#if userDetails.roles.length === 0}
+					<option selected>Ninguno</option>
+				{:else}
+					{#each userDetails.roles as { rol_id, rolDescription }}
+						<option value={rol_id}>{rolDescription}</option>
+					{/each}
+				{/if}
+			</select>
 		</div>
 	</div>
 </form>
